--- Koch/KochSnowflakeViewer.java
@@ -1,95 +1,0 @@
-//********************************************************************
-//  KochSnowflakeViewer.java       Author: Lewis/Loftus/Cocking
-//
-//  Demonstrates the use of recursion.
-//  @gcschmit (19 July 2014): converted from an applet to an application
-//********************************************************************
-
-import java.awt.*;
-import java.awt.event.*;
-import javax.swing.*;
-
-public class KochSnowflakeViewer implements ActionListener
-{
-    private final int WIDTH = 400;
-    private final int HEIGHT = 500;
-
-    private final int MIN = 1, MAX = 13;
-
-    private JButton increase, decrease;
-    private JLabel titleLabel, orderLabel;
-    private KochPanel drawing;
-    private JPanel panel, tools;
-    private JFrame frame;
-
-    //-----------------------------------------------------------------
-    //  Sets up the components for the applet.
-    //-----------------------------------------------------------------
-    public static void main(String[] args)
-    {
-        KochSnowflakeViewer viewer = new KochSnowflakeViewer();
-    }
-
-    public KochSnowflakeViewer()
-    {
-        tools = new JPanel ();
-        tools.setLayout (new BoxLayout(tools, BoxLayout.X_AXIS));
-        tools.setBackground (Color.yellow);
-        tools.setOpaque (true);
-
-        titleLabel = new JLabel ("The Koch Snowflake");
-        titleLabel.setForeground (Color.black);
-
-        increase = new JButton (new ImageIcon ("increase.gif"));
-        increase.setPressedIcon (new ImageIcon ("increasePressed.gif"));
-        increase.setMargin (new Insets (0, 0, 0, 0));
-        increase.addActionListener (this);
-        decrease = new JButton (new ImageIcon ("decrease.gif"));
-        decrease.setPressedIcon (new ImageIcon ("decreasePressed.gif"));
-        decrease.setMargin (new Insets (0, 0, 0, 0));
-        decrease.addActionListener (this);
-
-        orderLabel = new JLabel ("Order: 1");
-        orderLabel.setForeground (Color.black);
-
-        tools.add (titleLabel);
-        tools.add (Box.createHorizontalStrut (20));
-        tools.add (decrease);
-        tools.add (increase);
-        tools.add (Box.createHorizontalStrut (20));
-        tools.add (orderLabel);
-
-        drawing = new KochPanel (1);
-
-        panel = new JPanel();
-        panel.add (tools);
-        panel.add (drawing);
-
-        frame = new JFrame();
-        frame.setTitle("Knoch Snowflake");
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setSize(WIDTH, HEIGHT);
-        frame.add(panel);
-        frame.setVisible(true);
-    }
-    //-----------------------------------------------------------------
-    //  Determines which button was pushed, and sets the new order
-    //  if it is in range.
-    //-----------------------------------------------------------------
-    public void actionPerformed (ActionEvent event)
-    {
-        int order = drawing.getOrder();
-
-        if (event.getSource() == increase)
-            order++;
-        else
-            order--;
-
-        if (order >= MIN && order <= MAX)
-        {
-            orderLabel.setText ("Order: " + order);
-            drawing.setOrder (order);
-            frame.repaint();
-        }
-    }
-}
